visual: extract bit editor from vmenus.c to biteditor.c
