+++ librz/core/meson.build
@@ -101,6 +101,7 @@ rz_core_sources = [
   'cmd/cmd_regs.c',
   'cmd/cmd_linux_heap_glibc.c',
   'cmd/cmd_windows_heap.c',
+  'tui/biteditor.c',
   'tui/hud.c',
   'tui/visual.c',
   'tui/vmenus.c',
